moved countries and locationPostalCodes to data directory
